# backend/openapi.yaml
openapi: "3.1.0"
info:
  title: PaperFull API
  description: |
    AI-powered academic paper generation API. Authentication via httpOnly cookies
    + double-submit CSRF token (X-CSRF-TOKEN header for state-changing requests).
  version: "1.1.0"
  contact:
    name: PaperFull
    url: https://paperfull.app
servers:
  - url: https://paperfull.app
    description: Production
  - url: http://localhost:8000
    description: Local proxy (frontend)
  - url: http://localhost:8001
    description: Local backend (direct)
...
paths:
...
  /api/papers:
    get:
      tags: [Papers]
      summary: List papers (paginated)
      operationId: listPapers
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: Paginated list.
          content:
            application/json:
              schema:
                type: object
                required: [papers, pagination]
                properties:
                  papers:
                    type: array
                    items: { $ref: "#/components/schemas/PaperSummary" }
                  pagination: { $ref: "#/components/schemas/OffsetPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Papers]
      summary: Create or upsert a paper
      operationId: savePaper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string, pattern: "^[A-Za-z0-9_-]{1,64}$" }
                title: { type: string }
                data: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Saved.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  id: { type: string }
                  paper: { $ref: "#/components/schemas/PaperSummary" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List all users
      operationId: adminListUsers
      responses:
        "200":
          description: List of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    username:
                      type: string
                    email:
                      type: string
                    is_admin:
                      type: boolean
                    created_at:
                      type: string
                      format: date-time
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Forbidden (admin access required)
  /api/admin/papers:
    get:
      tags: [Admin]
      summary: List all papers across users
      operationId: adminListPapers
      responses:
        "200":
          description: List of all papers
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    title:
                      type: string
                    user_id:
                      type: integer
                    created_at:
                      type: string
                      format: date-time
                    status:
                      type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Forbidden (admin access required)
  /api/admin/usage:
    get:
      tags: [Admin]
      summary: API token usage stats
      operationId: adminUsage
      responses:
        "200":
          description: Usage statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_users:
                    type: integer
                  total_papers:
                    type: integer
                  total_generations:
                    type: integer
                  api_calls_today:
                    type: integer
                  storage_used_mb:
                    type: number
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Forbidden (admin access required)
  /api/admin/stats:
    get:
      tags: [Admin]
      summary: Summary counts
      operationId: adminStats
      responses:
        "200":
          description: System statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  uptime_seconds:
                    type: integer
                  active_users:
                    type: integer
                  queue_length:
                    type: integer
                  cache_hit_rate:
                    type: number
                  avg_response_time_ms:
                    type: number
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Forbidden (admin access required)
components:
...
  schemas:
...
    PaperSummary:
      type: object
      required: [id, title, user_id]
      properties:
        id: { type: string }
        title: { type: string }
        user_id: { type: integer }
        image_count: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    OffsetPage:
      type: object
      required: [limit, offset, total, has_more]
      properties:
        limit: { type: integer }
        offset: { type: integer }
        total: { type: integer }
        has_more: { type: boolean }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
      description: |
        httpOnly access cookie. CSRF protection: include the value of the
        `csrf_access_token` cookie as the `X-CSRF-TOKEN` header on all
        state-changing requests (POST/PUT/PATCH/DELETE).
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Legacy header-based auth (still accepted as fallback).
security:
  - cookieAuth: []
  - bearerAuth: []